[LB-419] refactor: 강의서버와 연동하며 요구사항에 맞게 코드 변경

+++ src/main/java/com/lgcms/consulting/domain/Progress.java
@@ -17,5 +17,5 @@ public class Progress {
 
     private String lectureId;
 
-    private Long progressRate;
+    private Integer progressRate;
 }
